METHODS RETURN ARRAYS
When a method returns an array, you can use it in a {{#method}}
section to iterate over the array.

Mustache::Simple->resolve

 - when in a section, call ->find in array context

Mustache::Simple::ContextStack->search

 - call the method in an array context
 - depending on wantarray(), return the array or the first element
 - CHANGE: return the actual value returned by the method call,
   not a coderef that will be executed later

I added a test file for the new functionality.

This commit passes all previous tests as well as the new tests,
but I can not help wondering if something about the change I made
will break backwards compatibility in a way that was not tested.